Fix for 64bit CPUs - use 32bit 'int' instead of 64bit 'long'

+++ ref-impl/src/OM/OMMSStructuredStorage.cpp
@@ -303,7 +303,7 @@ OMInt32 OMStgOpenStorageOnILockBytes(ILockBytes* plkbyt,
 
 //Below are the definititions for MS Strucutred Storage Library
 
-typedef unsigned long ULONG;    // 4 bytes
+typedef unsigned int ULONG;     // 4 bytes
 typedef unsigned short USHORT;  // 2 bytes
 typedef short OFFSET;           // 2 bytes
 typedef ULONG SECT;             // 4 bytes
@@ -313,8 +313,8 @@ typedef USHORT WCHAR;           // 2 bytes
 typedef ULONG DFSIGNATURE;      // 4 bytes
 typedef unsigned char BYTE;     // 1 byte
 typedef unsigned short WORD;    // 2 bytes
-typedef unsigned long DWORD;    // 4 bytes
-typedef ULONG StrID;              // 4 bytes
+typedef unsigned int DWORD;     // 4 bytes
+typedef ULONG StrID;            // 4 bytes
 
 //FAT Entries
 const SECT MAXREGSECT = 0xFFFFFFFA; // maximum SECT
